fix: (#196) usecase 로직 n+1 발생 안하게 변경 & 파일 이름 변경
